requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
httpx==0.26.0
python-jose[cryptography]==3.3.0
alembic==1.13.1
slowapi==0.1.9
python-json-logger==2.0.7
